feat(host-control): introduce HostControlHub abstraction
Add web/server/src/host_control/ as the foundation of the cross-mode
Tauri bridge unification (Step 1 of the host-control-hub plan):

- protocol.rs: unified HostControlMessage wire format with role
  filtering and serde forward-compat tolerance
- mod.rs: HostControlHub with Local/Aggregator/Forwarder modes;
  pending+replay+routing tables; fail-fast deny when no UI is reachable
- endpoint.rs: actix WebSocket handler with constant-time token auth
  and role-aware outbound filtering
- upstream.rs: forwarder ws-client scaffolding with exponential
  backoff (1s..30s capped)

37 unit tests cover U-1..U-17 of the implementation plan. Module is
self-contained; business call sites are wired up in subsequent steps.

SecurityPermissionType gains PartialEq + Eq so it can flow through the
new wire types.

Drive-by:
- cargo fmt --all picks up formatting drift across the workspace
- cargo clippy --fix --workspace + manual fixes in
  server/build.rs, signal/build.rs, utils/network.rs, and
  signal-facade/{desk_settings,signal,service}.rs

Approximately 50 historical clippy warnings remain in capture-engine
and input-injection; they are deferred to a separate chore commit.

+++ capture-engine/src/video_encoder/x264_encoder.rs
@@ -29,8 +29,8 @@ impl X264Encoder {
         display_info: &DisplayInfo,
         fps: u32,
     ) -> Result<Self, CaptureError> {
-        let width = display_info.desktop_coordinates.width() as i32;
-        let height = display_info.desktop_coordinates.height() as i32;
+        let width = display_info.desktop_coordinates.width();
+        let height = display_info.desktop_coordinates.height();
 
         let mut setup = Setup::preset(Preset::Ultrafast, Tune::None, false, true);
         setup = setup.crf(setting.quality as f32);
